ml-service-hf/requirements.txt
fastapi
uvicorn
dlib
Pillow
numpy
python-multipart
